server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
browser-use
openai